# FourSemester/Programming/Go/Tinkoff/2024-spring-go-course-lesson6-RustamNazarov2/dto/api/openapi.yaml
openapi: 3.0.1
info:
  description: Интерфейс управления и мониторинга устройствами умного дома
  title: API умного дома
  version: "0.1"
servers:
- url: http://localhost:8080/
...
paths:
...
  /sensors/{sensor_id}/events:
    get:
      description: Позволяет подписаться на рассылку последних событий пришедших от
        датчика
      parameters:
      - description: Идентификатор датчика
        in: path
        name: sensor_id
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "101":
          content: {}
          description: Успешное открытие ws
        default:
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
          description: Ошибка исполнения
      summary: Открытие ws по датчику
      tags:
      - sensors
...
components:
  schemas:
...
    Error:
      description: Ошибка исполнения запроса
      example:
        reason: Произошла ошибка
      properties:
        reason:
          description: Причина
          minLength: 1
          type: string
      required:
      - reason
      title: Error
      type: object
...